Combine roommanager and roompollingcomponent
